Moved impl + tests to identity.rs
